Weather-App

+++ app/src/main/java/com/example/weather/api/CurrentLocation.kt
@@ -2,7 +2,7 @@ package com.example.weather.api
 
 import com.squareup.moshi.Json
 
-data class Coordinates(val lat: Float, val lon: Float)
+data class Coordinates(val lat: Double, val lon: Double)
 
 data class CurrentLocation(
     val coord: Coordinates,
